Finished Timer
Timer now functions properly and counts down from the desired time down to 0 and then stops.

Timer has a setting to allow it to have a custom time to start from.

Timer has an empty function "_on_time_end" that has nothing in it yet that will be attached to future functionality to end minigames and move on to the correct next scene.